fixup! simplify signature helpers
